+++ packages/project/test/fixtures/sample-v1-project.ts
@@ -3,15 +3,15 @@ import { cloneDeep } from 'lodash-es';
 
 const state: Provisioner.Project = {
   id: 'e16c5f09-f0cb-4ba7-a4c2-73fcb2f29d00',
-  name: 'My Workflow',
+  name: 'My Project',
   description: 'a project',
   concurrency: null,
   inserted_at: '2025-04-23T11:15:59Z',
   collections: [],
   workflows: {
-    'my-workflow': {
+    'webhook-workflow': {
       id: '72ca3eb0-042c-47a0-a2a1-a545ed4a8406',
-      name: 'My Workflow',
+      name: 'Webhook Workflow',
       edges: {
         'trigger->transform-data': {
           enabled: true,
@@ -44,6 +44,44 @@ const state: Provisioner.Project = {
       lock_version: 1,
       deleted_at: null,
     },
+    'cron-workflow': {
+      id: 'd63934ea-6144-4dd3-b7a0-aa4e65d6129c',
+      name: 'Cron Workflow',
+      edges: {
+        'trigger->transform-data': {
+          enabled: true,
+          id: 'ee2b2cdf-ebd3-48f6-9176-488233fc00b5',
+          source_trigger_id: 'f2ebf159-d401-4b42-9815-143edb6a6ba0',
+          condition_type: 'always',
+          target_job_id: '8f30f10f-5e77-46ca-a603-c1fe70a2802e',
+        },
+      },
+      concurrency: null,
+      inserted_at: '2025-04-23T11:19:32Z',
+      updated_at: '2025-04-23T11:19:32Z',
+      jobs: {
+        'transform-data': {
+          id: 'e0c6bd0b-3320-4e5f-ac08-36d2ae2405bc',
+          name: 'Transform data',
+          body: 'fn(s => s)',
+          adaptor: '@openfn/language-common@latest',
+          project_credential_id: null,
+          keychain_credential_id: null,
+        },
+      },
+      triggers: {
+        cron: {
+          enabled: true,
+          id: '7ab78a65-81e7-4269-8aaf-ef70a80957ed',
+          type: 'cron',
+          cron_expression: '0 0 * * *',
+          // this must map correctly on the internal structure: in prod it fails
+          cron_cursor_job_id: 'e0c6bd0b-3320-4e5f-ac08-36d2ae2405bc',
+        },
+      },
+      lock_version: 1,
+      deleted_at: null,
+    },
   },
   updated_at: '2025-04-23T11:15:59Z',
   project_credentials: [],
@@ -66,7 +104,7 @@ withCreds.project_credentials = [
     owner: 'admin@openfn.org',
   },
 ];
-Object.assign(withCreds.workflows['my-workflow'].jobs['transform-data'], {
+Object.assign(withCreds.workflows['webhook-workflow'].jobs['transform-data'], {
   project_credential_id: 'p',
   keychain_credential_id: 'k',
 });